+++ test/test1.js
@@ -1,0 +1,14 @@
+<!doctype html>
+<html>
+<head>
+<meta http-equiv="Content-type" content="text/html; charset=utf-8" />
+</head>
+<body>
+</body>
+<script>
+avalon.define("blog", function(vm){
+    vm.name = 'hello';
+});
+</script>
+</body>
+</html>